Adds Erumu source code and dockerizes Erumu
This adds the Erumu DOM PureScript code and introduces a compose
file that builds and tests the application (though no test suite
currently exists).

Worth noting is that, in the `DOM.Erumu.HTML` re-export module,
all of the imports from the submodules had to be listed as
implicit imports to avoid raising the
`ImplicitQualifiedImportReExport` warning.