[sc-39347] Adds onPreventDefault
This adds a new event handler to prevent the default event (i.e.
navigating to the HREF of an anchor tag) from occurring.

This way we can have our cake and eat it to with links - supporting
links that both navigate using the single-page-app routing but also
support middle clicking or right click + copying the address.